[REFACTOR] refactor jwt and domain.report
