+++ src/main/java/com/sullung2yo/seatcatcher/common/domain/TokenType.java
@@ -1,0 +1,5 @@
+package com.sullung2yo.seatcatcher.common.domain;
+
+public enum TokenType {
+    ACCESS, REFRESH
+}
