+++ src/main/java/com/sullung2yo/seatcatcher/common/jwt/provider/JwtTokenProviderImpl.java
@@ -1,8 +1,8 @@
-package com.sullung2yo.seatcatcher.jwt.provider;
+package com.sullung2yo.seatcatcher.common.jwt.provider;
 
 import com.sullung2yo.seatcatcher.common.exception.ErrorCode;
 import com.sullung2yo.seatcatcher.common.exception.TokenException;
-import com.sullung2yo.seatcatcher.jwt.domain.TokenType;
+import com.sullung2yo.seatcatcher.common.domain.TokenType;
 import com.sullung2yo.seatcatcher.user.domain.RefreshToken;
 import com.sullung2yo.seatcatcher.user.domain.User;
 import com.sullung2yo.seatcatcher.user.domain.UserRole;
