+++ src/main/java/com/sullung2yo/seatcatcher/config/TestWebSocketConfig.java
@@ -1,6 +1,6 @@
 package com.sullung2yo.seatcatcher.config;
 
-import com.sullung2yo.seatcatcher.jwt.provider.TokenProvider;
+import com.sullung2yo.seatcatcher.common.jwt.provider.TokenProvider;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Configuration;
