+++ src/main/java/com/sullung2yo/seatcatcher/domain/report/entity/Report.java
@@ -1,6 +1,7 @@
-package com.sullung2yo.seatcatcher.user.domain;
+package com.sullung2yo.seatcatcher.domain.report.entity;
 
 import com.sullung2yo.seatcatcher.common.domain.BaseEntity;
+import com.sullung2yo.seatcatcher.user.domain.User;
 import jakarta.persistence.*;
 import lombok.*;
