+++ src/main/java/com/sullung2yo/seatcatcher/domain/report/repository/ReportRepository.java
@@ -1,6 +1,6 @@
-package com.sullung2yo.seatcatcher.user.repository;
+package com.sullung2yo.seatcatcher.domain.report.repository;
 
-import com.sullung2yo.seatcatcher.user.domain.Report;
+import com.sullung2yo.seatcatcher.domain.report.entity.Report;
 import com.sullung2yo.seatcatcher.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
