--- src/main/java/com/sullung2yo/seatcatcher/jwt/domain/TokenType.java
@@ -1,5 +1,0 @@
-package com.sullung2yo.seatcatcher.jwt.domain;
-
-public enum TokenType {
-    ACCESS, REFRESH
-}
