+++ src/test/java/com/sullung2yo/seatcatcher/user/controller/TagControllerTest.java
@@ -1,7 +1,7 @@
 package com.sullung2yo.seatcatcher.user.controller;
 
-import com.sullung2yo.seatcatcher.jwt.domain.TokenType;
-import com.sullung2yo.seatcatcher.jwt.provider.JwtTokenProviderImpl;
+import com.sullung2yo.seatcatcher.common.domain.TokenType;
+import com.sullung2yo.seatcatcher.common.jwt.provider.JwtTokenProviderImpl;
 import com.sullung2yo.seatcatcher.user.domain.*;
 import com.sullung2yo.seatcatcher.user.repository.UserRepository;
 import org.springframework.transaction.annotation.Transactional;
