+++ src/test/java/com/sullung2yo/seatcatcher/user/controller/TokenControllerTest.java
@@ -1,8 +1,8 @@
 package com.sullung2yo.seatcatcher.user.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.sullung2yo.seatcatcher.jwt.domain.TokenType;
-import com.sullung2yo.seatcatcher.jwt.provider.JwtTokenProviderImpl;
+import com.sullung2yo.seatcatcher.common.domain.TokenType;
+import com.sullung2yo.seatcatcher.common.jwt.provider.JwtTokenProviderImpl;
 import com.sullung2yo.seatcatcher.user.domain.Provider;
 import com.sullung2yo.seatcatcher.user.domain.User;
 
@@ -10,7 +10,6 @@ import com.sullung2yo.seatcatcher.user.dto.request.TokenRefreshRequest;
 import com.sullung2yo.seatcatcher.user.repository.RefreshTokenRepository;
 import com.sullung2yo.seatcatcher.user.repository.UserRepository;
 import org.springframework.transaction.annotation.Transactional;
-import org.aspectj.lang.annotation.After;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
