+++ src/test/java/com/sullung2yo/seatcatcher/user/service/UserAlarmServiceImplTest.java
@@ -1,8 +1,8 @@
 package com.sullung2yo.seatcatcher.user.service;
 
 import com.sullung2yo.seatcatcher.SeatCatcherApiApplication;
-import com.sullung2yo.seatcatcher.jwt.domain.TokenType;
-import com.sullung2yo.seatcatcher.jwt.provider.JwtTokenProviderImpl;
+import com.sullung2yo.seatcatcher.common.domain.TokenType;
+import com.sullung2yo.seatcatcher.common.jwt.provider.JwtTokenProviderImpl;
 import com.sullung2yo.seatcatcher.user.domain.*;
 import com.sullung2yo.seatcatcher.user.repository.UserAlarmRepository;
 import com.sullung2yo.seatcatcher.user.repository.UserRepository;
